requirements.txt
discum~=1.4.1
python-dotenv